Change the default search engine to searX
This commit changes the default search engine to searx, a privacy-respecting metasearch engine. Google are known for invasive privacy, limiting freedom, and censorship of information. I do not think it's wise to direct users to a dangerous search engine as a default feature. Instead, let's start them off with a safer option.